+++ CHANGELOG.md
@@ -23,6 +23,7 @@ This file is used to list changes made in each version of the openstack-dashboar
 * Allow webserver to be configurable
 * Fix site template directory defaults for apache 2.4
 * Use common specific_endpoint routines (bug 1412919)
+* Fix notify when using listen_addresses
 
 ## 9.1
 * python_packages database client attributes have been moved to the -common cookbook